backend/requirements.txt
fastapi
uvicorn
python-multipart
pythainlp
pytesseract
opencv-python-headless
Pillow
numpy